abstract copying package files over
